Ai_chatbot/requirements.txt
langchain
langchain-community
langchain-core
chromadb
streamlit
pandas
numpy
openai
python-dotenv
python-dateutil
langchain-openai
Ollama
